+++ src/pages/ThorChainLP/components/AddLiquitity/AddLiquidityConfirm.tsx
@@ -1,0 +1,3 @@
+export const AddLiquidityConfirm = () => {
+  return <p>Confirm</p>
+}
