+++ src/pages/ThorChainLP/components/AddLiquitity/types.ts
@@ -1,0 +1,4 @@
+export enum AddLiquidityRoutePaths {
+  Input = '/add/input',
+  Confirm = '/add/confirm',
+}
